Refactor statistics handling by replacing appendStats method with mergeStatistics in InstructorSessionResultPageComponent
