+++ src/web/app/components/question-types/question-statistics/question-statistics.ts
@@ -1,7 +1,5 @@
 import { Directive, Input } from '@angular/core';
 import {
-  ContributionStatistics,
-  ContributionStatisticsEntry,
   FeedbackParticipantType,
   FeedbackQuestionDetails,
   FeedbackResponseDetails,
@@ -39,29 +37,4 @@ export class QuestionStatistics<Q extends FeedbackQuestionDetails, R extends Fee
   constructor(question: Q) {
     this.question = question;
   }
-
-  static appendStats = (prevStats: string, newStats: string): string => {
-    if (prevStats === '') {
-      return newStats;
-    }
-    if (newStats === '') {
-      return prevStats;
-    }
-
-    // Stats not being empty means it belongs to contribution question
-    const prevStatsJSON: ContributionStatistics = JSON.parse(prevStats);
-    const newStatsJSON: ContributionStatistics = JSON.parse(newStats);
-    for (const email of Object.keys(newStatsJSON.results)) {
-      const newStatsEntryForEmail: ContributionStatisticsEntry = newStatsJSON.results[email];
-      const { claimed }: { claimed: number } = newStatsEntryForEmail;
-      const { perceived }: { perceived: number } = newStatsEntryForEmail;
-      if (claimed < 0 && perceived < 0) {
-        continue;
-      }
-      // If new entry has submitted stats, overwrite the old data
-      prevStatsJSON.results[email] = newStatsEntryForEmail;
-    }
-
-    return JSON.stringify(prevStatsJSON);
-  };
 }
